Move paymentsSlotClaimedExternally into the deposits domain
It gates the account-model deposit methods on whether DTCC/collateral
already claimed the single PaymentsPlugin slot — deposit-domain logic that
had no business in the generic registry.ts. Now a leaf module
(deposits/payments-slot.ts) consumed by the wallet/pull/ota registrars;
registry.ts is back to pure wiring.
