+++ src/integrations/registry.ts
@@ -10,12 +10,6 @@ import { registerDfns } from "./dfns";
 import { registerDeposits } from "./deposits";
 import { registerTokenStandards } from "./token-standards";
 
-/** True when another integration (DTCC, collateral, …) already owns the single PaymentsPlugin slot. */
-export function paymentsSlotClaimedExternally(): boolean {
-  return process.env.DTCC_PLUGIN_ENABLED === 'true'
-      || (!!process.env.COLLATERAL_REGISTRY_ADDRESS && !!process.env.COLLATERAL_AGENT_PRIVATE_KEY);
-}
-
 export interface IntegrationContext {
   orgId: string;
   logger: winston.Logger;
